ENH/MNT/REF: upstream FilestoreBackend.compare to _Backend, implement .search and .root for TestBackend
